Start from chapter X if provided; Improve Cleanup()
If a chapter or page URL is provided, as if we should start at the
given chapter instead of the beginning chapter.  If yes, do not process
chapters before the one given.

Improve Cleanup() a bit by removing empty chapter folders.  Also
silently ignore "file not found" errors when trying to delete them.